Saving/loading of active config will be skipped, when active_config setting is empty Loading of default config will be skipped, when default_config setting is empty

+++ filemanagement.py
@@ -89,6 +89,8 @@ def get_active_config(active_config):
 
 
 def set_as_active_config(active_config, file):
+    if not active_config:
+        return
     if islink(active_config):
         os.unlink(active_config)
     os.symlink(file, active_config)
